+++ client/src/main/java/com/google/cloud/gcs/analyticscore/client/GcsClientImpl.java
@@ -26,10 +26,7 @@ import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobId;
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.BlobWriteSession;
-import com.google.cloud.storage.BlobWriteSessionConfig;
-import com.google.cloud.storage.BlobWriteSessionConfigs;
 import com.google.cloud.storage.HttpStorageOptions;
-import com.google.cloud.storage.ParallelCompositeUploadBlobWriteSessionConfig;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.Storage.BlobWriteOption;
 import com.google.cloud.storage.StorageException;
@@ -40,14 +37,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
 import java.nio.channels.WritableByteChannel;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.ExecutorService;
-import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -125,14 +117,15 @@ class GcsClientImpl implements GcsClient {
   }
 
   @Override
-  public WritableByteChannel create(GcsItemId itemId, GcsWriteOptions writeOptions)
+  public WritableByteChannel createWriteChannel(GcsItemId itemId, GcsWriteOptions writeOptions)
       throws IOException {
     checkNotNull(itemId, "itemId should not be null");
 
     BlobInfo blobInfo = createBlobInfo(itemId);
 
     try {
-      BlobWriteOption[] sdkWriteOptions = generateWriteOptions(writeOptions, itemId);
+      BlobWriteOption[] sdkWriteOptions =
+          GcsWriteConfigurationUtil.generateWriteOptions(writeOptions, itemId);
       BlobWriteSession sdkWriteSession = storage.blobWriteSession(blobInfo, sdkWriteOptions);
       WritableByteChannel channel = sdkWriteSession.open();
       return new GcsWriteChannel(sdkWriteSession, channel, blobInfo, writeOptions);
@@ -140,107 +133,6 @@ class GcsClientImpl implements GcsClient {
       throw GcsExceptionUtil.translateWriteException(
           e, "initialization", blobInfo.getBlobId(), 0L, writeOptions);
     }
-  }
-
-  private BlobWriteSessionConfig generateSessionConfig(
-      GcsWriteOptions writeOptions, boolean isHttpTransport) throws IOException {
-    switch (writeOptions.getUploadType()) {
-      case PARALLEL_COMPOSITE_UPLOAD:
-        return getParallelCompositeUploadSessionConfig(writeOptions);
-      case WRITE_TO_DISK_THEN_UPLOAD:
-        return getWriteToDiskSessionConfig(writeOptions);
-      case JOURNALING:
-        return getJournalingSessionConfig(writeOptions, isHttpTransport);
-      case CHUNK_UPLOAD:
-        return BlobWriteSessionConfigs.getDefault()
-            .withChunkSize(writeOptions.getUploadChunkSize());
-      default:
-        return BlobWriteSessionConfigs.getDefault();
-    }
-  }
-
-  private BlobWriteSessionConfig getParallelCompositeUploadSessionConfig(
-      GcsWriteOptions writeOptions) {
-    return BlobWriteSessionConfigs.parallelCompositeUpload()
-        .withBufferAllocationStrategy(
-            ParallelCompositeUploadBlobWriteSessionConfig.BufferAllocationStrategy.fixedPool(
-                writeOptions.getPcuBufferCount(), writeOptions.getPcuBufferCapacity()))
-        .withPartCleanupStrategy(getSdkCleanupStrategy(writeOptions.getPcuPartFileCleanupType()))
-        .withPartNamingStrategy(
-            ParallelCompositeUploadBlobWriteSessionConfig.PartNamingStrategy.prefix(
-                writeOptions.getPcuPartFileNamePrefix()));
-  }
-
-  private BlobWriteSessionConfig getWriteToDiskSessionConfig(GcsWriteOptions writeOptions)
-      throws IOException {
-    if (!writeOptions.getTemporaryPaths().isEmpty()) {
-      List<Path> paths = toPaths(writeOptions.getTemporaryPaths());
-      return BlobWriteSessionConfigs.bufferToDiskThenUpload(paths);
-    } else {
-      return BlobWriteSessionConfigs.bufferToTempDirThenUpload();
-    }
-  }
-
-  private BlobWriteSessionConfig getJournalingSessionConfig(
-      GcsWriteOptions writeOptions, boolean isHttpTransport) throws IOException {
-    if (isHttpTransport) {
-      throw new UnsupportedOperationException(
-          "JOURNALING upload type is not supported because it requires the gRPC "
-              + "transport backend (HTTP transport is currently active).");
-    }
-    checkArgument(
-        !writeOptions.getTemporaryPaths().isEmpty(),
-        "Temporary paths must be configured for JOURNALING upload type");
-    List<Path> paths = toPaths(writeOptions.getTemporaryPaths());
-    return BlobWriteSessionConfigs.journaling(paths);
-  }
-
-  private static List<Path> toPaths(Collection<String> pathStrings) {
-    return pathStrings.stream().map(Paths::get).collect(Collectors.toList());
-  }
-
-  private ParallelCompositeUploadBlobWriteSessionConfig.PartCleanupStrategy getSdkCleanupStrategy(
-      GcsWriteOptions.PartFileCleanupType cleanupType) {
-    switch (cleanupType) {
-      case NEVER:
-        return ParallelCompositeUploadBlobWriteSessionConfig.PartCleanupStrategy.never();
-      case ON_SUCCESS:
-        return ParallelCompositeUploadBlobWriteSessionConfig.PartCleanupStrategy.onlyOnSuccess();
-      case ALWAYS:
-      default:
-        return ParallelCompositeUploadBlobWriteSessionConfig.PartCleanupStrategy.always();
-    }
-  }
-
-  private BlobWriteOption[] generateWriteOptions(GcsWriteOptions writeOptions, GcsItemId itemId) {
-    List<BlobWriteOption> sdkWriteOptions = new ArrayList<>();
-
-    if (writeOptions != null) {
-      if (writeOptions.isDisableGzipContent()) {
-        sdkWriteOptions.add(BlobWriteOption.disableGzipContent());
-      }
-      if (writeOptions.isChecksumValidationEnabled()) {
-        sdkWriteOptions.add(BlobWriteOption.crc32cMatch());
-      }
-      writeOptions.getKmsKeyName().map(BlobWriteOption::kmsKeyName).ifPresent(sdkWriteOptions::add);
-      writeOptions
-          .getEncryptionKey()
-          .map(BlobWriteOption::encryptionKey)
-          .ifPresent(sdkWriteOptions::add);
-      writeOptions
-          .getUserProject()
-          .map(BlobWriteOption::userProject)
-          .ifPresent(sdkWriteOptions::add);
-    }
-
-    // Determine overwrite semantics based on exact generation ID or 'doesNotExist' flag
-    if (itemId.getContentGeneration().isPresent()) {
-      sdkWriteOptions.add(BlobWriteOption.generationMatch());
-    } else if (writeOptions != null && !writeOptions.isOverwriteExisting()) {
-      sdkWriteOptions.add(BlobWriteOption.doesNotExist());
-    }
-
-    return sdkWriteOptions.toArray(new BlobWriteOption[0]);
   }
 
   @Override
@@ -277,7 +169,7 @@ class GcsClientImpl implements GcsClient {
 
     boolean isHttp = builder instanceof HttpStorageOptions.Builder;
     builder.setBlobWriteSessionConfig(
-        generateSessionConfig(clientOptions.getGcsWriteOptions(), isHttp));
+        GcsWriteConfigurationUtil.generateSessionConfig(clientOptions, isHttp));
 
     return builder.build().getService();
   }
@@ -326,6 +218,7 @@ class GcsClientImpl implements GcsClient {
   }
 
   private BlobInfo createBlobInfo(GcsItemId itemId) {
+    checkNotNull(itemId, "itemId should not be null");
     BlobId blobId;
     if (itemId.getContentGeneration().isPresent()) {
       blobId =
