+++ client/src/test/java/com/google/cloud/gcs/analyticscore/client/GcsWriteOptionsTest.java
@@ -17,16 +17,12 @@
 package com.google.cloud.gcs.analyticscore.client;
 
 import static com.google.common.truth.Truth.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 class GcsWriteOptionsTest {
-
-  private static final int MB = 1024 * 1024;
 
   @Test
   void builder_withDefaultValues_returnsExpectedDefaults() {
@@ -35,14 +31,6 @@ class GcsWriteOptionsTest {
     assertThat(options.isChecksumValidationEnabled()).isFalse();
     assertThat(options.isDisableGzipContent()).isTrue();
     assertThat(options.isOverwriteExisting()).isTrue();
-    assertThat(options.getUploadChunkSize()).isEqualTo(24 * MB);
-    assertThat(options.getUploadType()).isEqualTo(GcsWriteOptions.UploadType.CHUNK_UPLOAD);
-    assertThat(options.getPcuBufferCount()).isEqualTo(1);
-    assertThat(options.getPcuBufferCapacity()).isEqualTo(32 * MB);
-    assertThat(options.getPcuPartFileCleanupType())
-        .isEqualTo(GcsWriteOptions.PartFileCleanupType.ALWAYS);
-    assertThat(options.getPcuPartFileNamePrefix()).isEmpty();
-    assertThat(options.getTemporaryPaths()).isEmpty();
     assertThat(options.getKmsKeyName().isPresent()).isFalse();
     assertThat(options.getUserProject().isPresent()).isFalse();
     assertThat(options.getEncryptionKey().isPresent()).isFalse();
@@ -55,13 +43,6 @@ class GcsWriteOptionsTest {
             .setChecksumValidationEnabled(true)
             .setDisableGzipContent(false)
             .setOverwriteExisting(false)
-            .setUploadChunkSize(1024)
-            .setUploadType(GcsWriteOptions.UploadType.PARALLEL_COMPOSITE_UPLOAD)
-            .setPcuBufferCount(4)
-            .setPcuBufferCapacity(64 * MB)
-            .setPcuPartFileCleanupType(GcsWriteOptions.PartFileCleanupType.ON_SUCCESS)
-            .setPcuPartFileNamePrefix("temp-prefix-")
-            .setTemporaryPaths(ImmutableList.of("/tmp/path1", "/tmp/path2"))
             .setKmsKeyName("kms-key")
             .setUserProject("project-123")
             .setEncryptionKey("enc-key")
@@ -70,15 +51,6 @@ class GcsWriteOptionsTest {
     assertThat(options.isChecksumValidationEnabled()).isTrue();
     assertThat(options.isDisableGzipContent()).isFalse();
     assertThat(options.isOverwriteExisting()).isFalse();
-    assertThat(options.getUploadChunkSize()).isEqualTo(1024);
-    assertThat(options.getUploadType())
-        .isEqualTo(GcsWriteOptions.UploadType.PARALLEL_COMPOSITE_UPLOAD);
-    assertThat(options.getPcuBufferCount()).isEqualTo(4);
-    assertThat(options.getPcuBufferCapacity()).isEqualTo(64 * MB);
-    assertThat(options.getPcuPartFileCleanupType())
-        .isEqualTo(GcsWriteOptions.PartFileCleanupType.ON_SUCCESS);
-    assertThat(options.getPcuPartFileNamePrefix()).isEqualTo("temp-prefix-");
-    assertThat(options.getTemporaryPaths()).containsExactly("/tmp/path1", "/tmp/path2").inOrder();
     assertThat(options.getKmsKeyName()).hasValue("kms-key");
     assertThat(options.getUserProject()).hasValue("project-123");
     assertThat(options.getEncryptionKey()).hasValue("enc-key");
@@ -91,13 +63,6 @@ class GcsWriteOptionsTest {
             .put("gcs.channel.write.checksum-validation.enabled", "true")
             .put("gcs.channel.write.disable-gzip-content", "false")
             .put("gcs.channel.write.overwrite-existing", "false")
-            .put("gcs.channel.write.chunk-size-bytes", "1024")
-            .put("gcs.channel.write.upload-type", "parallel_composite_upload")
-            .put("gcs.channel.write.pcu.buffer.count", "4")
-            .put("gcs.channel.write.pcu.buffer.capacity-bytes", "67108864")
-            .put("gcs.channel.write.pcu.part-file.cleanup-type", "on_success")
-            .put("gcs.channel.write.pcu.part-file.name-prefix", "temp-prefix-")
-            .put("gcs.channel.write.temporary-paths", "/tmp/path1, /tmp/path2")
             .put("gcs.kms-key-name", "kms-key")
             .put("gcs.user-project", "project-123")
             .put("gcs.encryption-key", "enc-key")
@@ -108,74 +73,8 @@ class GcsWriteOptionsTest {
     assertThat(options.isChecksumValidationEnabled()).isTrue();
     assertThat(options.isDisableGzipContent()).isFalse();
     assertThat(options.isOverwriteExisting()).isFalse();
-    assertThat(options.getUploadChunkSize()).isEqualTo(1024);
-    assertThat(options.getUploadType())
-        .isEqualTo(GcsWriteOptions.UploadType.PARALLEL_COMPOSITE_UPLOAD);
-    assertThat(options.getPcuBufferCount()).isEqualTo(4);
-    assertThat(options.getPcuBufferCapacity()).isEqualTo(64 * MB);
-    assertThat(options.getPcuPartFileCleanupType())
-        .isEqualTo(GcsWriteOptions.PartFileCleanupType.ON_SUCCESS);
-    assertThat(options.getPcuPartFileNamePrefix()).isEqualTo("temp-prefix-");
-    assertThat(options.getTemporaryPaths()).containsExactly("/tmp/path1", "/tmp/path2").inOrder();
     assertThat(options.getKmsKeyName()).hasValue("kms-key");
     assertThat(options.getUserProject()).hasValue("project-123");
     assertThat(options.getEncryptionKey()).hasValue("enc-key");
-  }
-
-  @Test
-  void createFromOptions_withWhitespaceAndEmptyPaths_filtersThemOut() {
-    Map<String, String> rawOptions =
-        ImmutableMap.of("gcs.channel.write.temporary-paths", "  , /tmp/path1 , , /tmp/path2 ");
-    GcsWriteOptions options = GcsWriteOptions.createFromOptions(rawOptions, "gcs.");
-    assertThat(options.getTemporaryPaths()).containsExactly("/tmp/path1", "/tmp/path2").inOrder();
-  }
-
-  @Test
-  void createFromOptions_withHyphenatedEnums_normalizesAndParsesCorrectly() {
-    Map<String, String> rawOptions =
-        ImmutableMap.<String, String>builder()
-            .put("gcs.channel.write.upload-type", "parallel-composite-upload")
-            .put("gcs.channel.write.pcu.part-file.cleanup-type", "on-success")
-            .build();
-
-    GcsWriteOptions options = GcsWriteOptions.createFromOptions(rawOptions, "gcs.");
-
-    assertThat(options.getUploadType())
-        .isEqualTo(GcsWriteOptions.UploadType.PARALLEL_COMPOSITE_UPLOAD);
-    assertThat(options.getPcuPartFileCleanupType())
-        .isEqualTo(GcsWriteOptions.PartFileCleanupType.ON_SUCCESS);
-  }
-
-  @Test
-  void createFromOptions_withOverflowingUploadChunkSize_throwsIllegalArgumentException() {
-    assertOverflowThrows("gcs.channel.write.chunk-size-bytes");
-  }
-
-  @Test
-  void createFromOptions_withOverflowingPcuBufferCount_throwsIllegalArgumentException() {
-    assertOverflowThrows("gcs.channel.write.pcu.buffer.count");
-  }
-
-  @Test
-  void createFromOptions_withOverflowingPcuBufferCapacity_throwsIllegalArgumentException() {
-    assertOverflowThrows("gcs.channel.write.pcu.buffer.capacity-bytes");
-  }
-
-  private void assertOverflowThrows(String key) {
-    Map<String, String> rawOptions =
-        ImmutableMap.of(key, String.valueOf((long) Integer.MAX_VALUE + 1L));
-    IllegalArgumentException exception =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> GcsWriteOptions.createFromOptions(rawOptions, "gcs."));
-    assertThat(exception.getMessage()).contains(key);
-    assertThat(exception.getMessage()).contains("cannot be greater than Integer.MAX_VALUE");
-  }
-
-  @Test
-  void createFromOptions_withOnlyWhitespacePaths_doesNotSetPaths() {
-    Map<String, String> rawOptions = ImmutableMap.of("gcs.channel.write.temporary-paths", "   ");
-    GcsWriteOptions options = GcsWriteOptions.createFromOptions(rawOptions, "gcs.");
-    assertThat(options.getTemporaryPaths()).isEmpty();
   }
 }
